feat: add MultiImageUpload component for handling multiple image uploads with drag-and-drop support
feat: enhance Select component to support required field indication

fix: update SortableTh component to set default props for sortBy and sortOrder

feat: improve TextField component to indicate required fields and show validation errors

feat: add book-related labels to constants for better localization

feat: implement BookManagementPage for managing book records with search, pagination, and modal forms

feat: create book service for API interactions related to book management

feat: extend category service to fetch child categories

feat: add format utility for price and number formatting

feat: enhance validation utility to include money validation logic

+++ back/src/main/resources/db/migration/V1.0.0__create_master_table.sql
@@ -123,7 +123,7 @@ CREATE TABLE public.bs_books (
     "layout" varchar(255),
     "name" varchar(255) NOT NULL,
     "page_count" int4,
-    "price" int8 NOT NULL,
+    "price" numeric(12,0) NOT NULL,
     "publish_year" int4,
     "quantity" int4 NOT NULL DEFAULT 0,
     "size" varchar(255),
